backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
groq
psycopg2-binary
pytz
requests